refactor: move Config to a separate module
